refactor: chat 도메인 관련 schema를 entities 레이어로 이동 (#69)
