+++ src/entities/message/model/schema.ts
@@ -1,47 +1,5 @@
 import z from 'zod'
 
-// ---------- 채팅방 목록 조회 ----------
-export const ChatRoomSchema = z
-  .object({
-    id: z.number(),
-    name: z.string(),
-    description: z.string(),
-    participant_count: z.number(),
-    last_message_at: z.coerce.date(),
-    created_at: z.coerce.date(),
-  })
-  .transform((data) => ({
-    id: data.id,
-    name: data.name,
-    description: data.description,
-    participantCount: data.participant_count,
-    lastMessageAt: data.last_message_at,
-    createdAt: data.created_at,
-  }))
-
-export type ChatRoom = z.infer<typeof ChatRoomSchema>
-
-export const ChatRoomListResponseSchema = z
-  .object({
-    rooms: z.array(ChatRoomSchema),
-  })
-  .transform((data) => ({
-    rooms: data.rooms,
-  }))
-
-export type ChatRoomListResponse = z.infer<typeof ChatRoomListResponseSchema>
-
-// ---------- 채팅방 입장 ----------
-export const ChatRoomEnterResponseSchema = z.object({
-  message: z.string(),
-  room: z.object({
-    id: z.number(),
-    name: z.string(),
-  }),
-})
-
-export type ChatRoomEnterResponse = z.infer<typeof ChatRoomEnterResponseSchema>
-
 // ---------- 채팅 메세지 조회 ----------
 export const ChatMessageListRequestSchema = z.object({
   roomId: z.number(),
@@ -104,10 +62,3 @@ export const ChatMessageListResponseSchema = z
 export type ChatMessageListResponse = z.infer<
   typeof ChatMessageListResponseSchema
 >
-
-// ---------- 실패 ----------
-export const ChatErrorResponseSchema = z.object({
-  detail: z.string(),
-})
-
-export type ChatErrorResponse = z.infer<typeof ChatErrorResponseSchema>
